+++ client/src/pages/MyPage/components/SectionArtistCardList/SectionArtistCardList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './SectionArtistCardList.module.css';
-import SectionArtistCard from '../SectionAtristCard/SectionArtistCard';
+import SectionArtistCard from '../Sections/SectionAtristCard/SectionArtistCard';
 
 export default function SectionArtistCardList({ items }) {
   return (
