+++ client/src/pages/MyPage/components/Sections/SectionAtristCard/SectionArtistCard.jsx
