+++ client/src/pages/MyPage/components/Sections/SectionCard/SectionCard.jsx
