+++ client/src/pages/MyPage/components/Sections/SectionCatalog/SectionCatalog.module.css
