+++ client/src/pages/MyPage/components/TabContent/TabLike/TabLike.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styles from './TabLike.module.css';
-import dummyImg from '../../../../assets/images/dummyImg.png';
-import SectionTitle from '../SectionTitle/SectionTitle';
-import SectionCardList from '../SectionCardList/SectionCardList';
-import SectionArtistCardList from '../SectionArtistCardList/SectionArtistCardList';
+import dummyImg from '../../../../../assets/images/dummyImg.png';
+import SectionTitle from '../../SectionTitle/SectionTitle';
+import SectionCardList from '../../SectionCardList/SectionCardList';
+import SectionArtistCardList from '../../SectionArtistCardList/SectionArtistCardList';
 
 const dummyBookmarks = [
   {
